requirements.txt
Flask==2.0.2
tensorflow==2.5.0
keras==2.4.3
pillow==8.2.0
numpy==1.19.5
opencv-python==4.5.2.54